Switch to using a CookieStore instead of a BoltDB store
This is (or something similar) is required if there are multiple splunk
indexers as the state can't be saved to a single node.

To deploy you'll need to set a SESSION_ENCRYPTION_KEY first. It may also
be wise to change the COOKIE_NAME as I'm not sure if we'll be able to
use the old cookie that boltdb used.

This is untested, but should work.